Refactor Firebase configuration into components/firebase directory
- Move FirebaseConfig.js to components/firebase/ for better organization
- Update import paths in GoogleSignUp.js to use new firebase module structure
- Separate auth functions from main firebase config
